fix base linux build, adding freetype support
git-svn-id: http://skia.googlecode.com/svn/trunk@371 2bbb7eff-a529-9590-31e7-b0007b416f81

+++ src/utils/SkSfntUtils.cpp
@@ -51,7 +51,7 @@ bool SkSfntUtils::ReadTable_head(SkFontID fontID, SkSfntTable_head* head) {
     head->fFontDirectionHint = parse_be16(p);
     head->fIndexToLocFormat = parse_be16(p);
     head->fGlyphDataFormat = parse_be16(p);
-    SkASSERT(p - storage == size);
+    SkASSERT(p - storage == (long)size);
     return true;
 }
 
@@ -81,7 +81,7 @@ bool SkSfntUtils::ReadTable_maxp(SkFontID fontID, SkSfntTable_maxp* maxp) {
     maxp->fMaxSizeOfInstructions = parse_be16(p);
     maxp->fMaxComponentElements = parse_be16(p);
     maxp->fMaxComponentDepth = parse_be16(p);
-    SkASSERT(p - storage == size);
+    SkASSERT(p - storage == (long)size);
     return true;
 }
